moved baseclass
